+++ CHANGELOG.md
@@ -18,6 +18,7 @@ under development
 - Add `Context.RouteURL` to generate URL of the naming route.
 - Add `Context.Emit` and `Context.Blob`.
 - Add `Context.HtmlBlob`.
+- Add `Context.XMLBlob`.
 
 v1.8.1 April 2, 2020
 --------------------